add reusable methods into driver
