+++ .idea/workspace.xml
@@ -2,9 +2,6 @@
 <project version="4">
   <component name="ChangeListManager">
     <list default="true" id="e318267c-f610-47c6-b0af-23d8972a7b06" name="Default Changelist" comment="">
-      <change afterPath="$PROJECT_DIR$/src/test/java/gmibank_team06/pojos/LoginPojo.java" afterDir="false" />
-      <change afterPath="$PROJECT_DIR$/src/test/java/gmibank_team06/runners/Runner.java" afterDir="false" />
-      <change afterPath="$PROJECT_DIR$/src/test/java/gmibank_team06/step_definitions/LoginStepDefinitions.java" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
     </list>
     <option name="SHOW_DIALOG" value="false" />
@@ -65,10 +62,10 @@
       <screen x="0" y="0" width="1536" height="824" />
     </state>
     <state x="424" y="5" key="CommitChangelistDialog2/0.0.1536.824@0.0.1536.824" timestamp="1604675358432" />
-    <state x="465" y="236" key="com.intellij.ide.util.TipDialog" timestamp="1604758403887">
+    <state x="465" y="236" key="com.intellij.ide.util.TipDialog" timestamp="1604786619817">
       <screen x="0" y="0" width="1536" height="824" />
     </state>
-    <state x="465" y="236" key="com.intellij.ide.util.TipDialog/0.0.1536.824@0.0.1536.824" timestamp="1604758403887" />
+    <state x="465" y="236" key="com.intellij.ide.util.TipDialog/0.0.1536.824@0.0.1536.824" timestamp="1604786619817" />
     <state x="578" y="329" key="com.intellij.openapi.vcs.update.UpdateOrStatusOptionsDialogupdate-v2" timestamp="1604672559730">
       <screen x="0" y="0" width="1536" height="824" />
     </state>
